Moved Command and Window classes to separate files.
